actually fix dtype

+++ influence_benchmark/RL/run_KTO_iteration.py
@@ -59,7 +59,7 @@ def train_kto():
     if kto_config.seed is not None:
         set_all_seeds(kto_config.seed)
 
-    tokenizer = AutoTokenizer.from_pretrained(args.model_name, dtype=torch.bfloat16)
+    tokenizer = AutoTokenizer.from_pretrained(args.model_name)
 
     def format_dataset(example):
         if "gemma" in args.model_name:
@@ -96,7 +96,7 @@ def train_kto():
     dataset = dataset.shuffle()  # type: ignore
     dataset = dataset.map(format_dataset, batched=False)
 
-    model = AutoModelForCausalLM.from_pretrained(args.model_name)
+    model = AutoModelForCausalLM.from_pretrained(args.model_name, dtype=torch.bfloat16)
     model.config.use_cache = False
     if getattr(model.config, "pad_token_id", None) is None:
         if "Llama-3.1" in args.model_name:
